requirements.txt
streamlit
numpy
Pillow
streamlit_option_menu